refactor: move icons to separate folder
