+++ packages/curve-common/src/widgets/Header/SocialSidebarSection.tsx
@@ -5,8 +5,8 @@ import TwitterIcon from '@mui/icons-material/Twitter'
 import YouTubeIcon from '@mui/icons-material/YouTube'
 import Tooltip from '@mui/material/Tooltip'
 import Link from '@mui/material/Link'
-import { DiscordIcon } from 'curve-ui-kit/src/shared/ui/DiscordIcon'
-import { DodoIcon } from 'curve-ui-kit/src/shared/ui/DodoIcon'
+import { DiscordIcon } from 'curve-ui-kit/src/shared/icons/DiscordIcon'
+import { DodoIcon } from 'curve-ui-kit/src/shared/icons/DodoIcon'
 import Box from '@mui/material/Box'
 import SvgIcon from '@mui/material/SvgIcon'
 import type { Locale } from 'curve-common/src/widgets/Header/types'
