+++ packages/curve-ui-kit/src/shared/icons/CrvHubIcon.tsx
