+++ packages/curve-ui-kit/src/shared/icons/DiscordIcon.tsx
