+++ packages/curve-ui-kit/src/shared/icons/DodoIcon.tsx
