+++ packages/curve-ui-kit/src/shared/icons/LAFIcon.tsx
