+++ packages/curve-ui-kit/src/shared/icons/XIcon.tsx
